requirements.txt
Django==2.2.7
psycopg2-binary==2.8.4
Pillow==6.2.1
django-resized==0.3.9
gunicorn==20.0.4
six==1.13.0
django-modeltranslation==0.14
